Changed name of abstract class, abstracted more
